Add files via upload
This dataset contained 198 observations from 92 companies belonging on the Nasdaq 100, were each observation represents the number times a company was mentioned on Twitter on a given day of trading from June 21, 2012 to September 18, 2013. From this data, we wanted to use different methods in order to predict links and therefore build an undirected network between these 92 Nasdaq companies. We will link companies which have a statistically significant correlation while using the number of tweets mentioning them on each day.  More specifically, given a set of observations of the Nasdaq 100 companies, the task of inferring this association network graph, is the same as inferring the set of correlations which are non-zero.